+++ src/soulstruct/havok/fromsoft/base/__init__.py
@@ -12,6 +12,7 @@ __all__ = [
     "BaseRagdollHKX",
     "SkeletonMapper",
     "BaseRemoAnimationHKX",
+    "RemoPartTracks",
     "BaseANIBND",
 ]
 
@@ -20,5 +21,5 @@ from .animation import BaseAnimationHKX, AnimationContainer
 from .physics import BaseCollisionHKX, BaseClothHKX, PhysicsData, ClothPhysicsData
 from .skeleton import BaseSkeletonHKX, Skeleton, Bone
 from .ragdoll import BaseRagdollHKX, SkeletonMapper
-from .remo_animation import BaseRemoAnimationHKX
+from .remo_animation import BaseRemoAnimationHKX, RemoPartTracks
 from .anibnd import BaseANIBND