+++ src/soulstruct/havok/fromsoft/base/remo_animation/__init__.py
@@ -1,3 +1,3 @@
-__all__ = ["BaseRemoAnimationHKX"]
+__all__ = ["BaseRemoAnimationHKX", "RemoPartTracks"]
 
-from .core import BaseRemoAnimationHKX
+from .core import BaseRemoAnimationHKX, RemoPartTracks